Perfect example for CoreData

+++ Bookwormv2/Views/BookRatingView.swift
@@ -7,7 +7,7 @@
 
 import SwiftUI
 struct BookRatingView: View {
-   @Binding var rating: Int
+   @Binding var rating: Int16
    
    var label = ""
    
@@ -30,7 +30,7 @@ struct BookRatingView: View {
             image(for: number)
                .foregroundColor(number > rating ? offColor : onColor)
                .onTapGesture {
-                  rating = number
+                  rating = Int16(number)
                }
          }
       }
